Share sample-rate normalization with TraceProvider.get_service()
PR #4371 (issue #4234, Langfuse trace sampling) merged into main during
this session, landing the exact sample_rate=0.0/blank normalization logic
in TraceProvider.get_service() that apps/help/tracing.py had its own
independent copy of. Extracted into
apps.service_providers.tracing.langfuse.normalize_sample_rate(), used by
both call sites now. No behavior change: the existing
TestTraceProviderGetService suite (apps/service_providers/tests/
test_trace_provider_sample_rate.py) passes unchanged, and the help-bot
side keeps its own test coverage via the shared function's new direct
test.
